pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gleb Buzin", email = "qufiwefefwoyn@gmail.com"},
]
dependencies = [
    "pyserial>=3.5",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
